backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic 
requests
PyJWT
python-multipart
passlib[bcrypt]
email-validator